fix(api): move more auth endpoints to new format
